+++ addons/godot_dotnet_mcp/plugin/runtime/plugin_runtime_state.gd
@@ -43,7 +43,6 @@ const DEFAULT_COLLAPSED_SYSTEM_TOOLS: Array[String] = [
 	"system_project_index_build",
 	"system_project_run",
 	"system_project_state",
-	"system_project_stop",
 	"system_project_symbol_search",
 	"system_runtime_control",
 	"system_runtime_diagnose",